Fix dashboard crash, support ticket error, and full UI redesign
Bug fixes:
- DashboardController: cast query-builder sum() to int (PHP returns
  string "0", which caused a Dart TypeError crashing the dashboard)
- SupportMessage: add is_staff to fillable (was causing 500 on ticket
  creation/reply)

Full UI redesign (professional, colorful):
- New theme with gradients (green, blue, purple, orange), card shadows
- Dashboard: gradient hero header with clicks today, colorful stat
  cards with icons, improved chart with gradient fill, contract card
- Stats: animated period filter chips, metric cards with icons/gradients
- Payments: redesigned balance cards, network selector with gradients,
  withdrawal card with status indicators
- Support: modern ticket list with colored icons, category chips in
  create form, gradient chat bubbles in detail screen
- Profile: gradient profile header card, improved forms with icons
- Login: two-tone design with gradient top + white bottom sheet
- Splash: full gradient background
- Widgets: stat cards with icon support, gradient-capable BalanceCard,
  improved error widget

+++ app/Http/Controllers/Api/DashboardController.php
@@ -19,8 +19,8 @@ class DashboardController extends Controller
 
         $stats = [
             'clicks_today'      => $todayEarning?->valid_clicks ?? 0,
-            'clicks_this_week'  => DailyEarning::where('user_id', $user->id)->whereBetween('date', [now()->startOfWeek(), now()])->sum('valid_clicks'),
-            'clicks_this_month' => DailyEarning::where('user_id', $user->id)->whereMonth('date', now()->month)->whereYear('date', now()->year)->sum('valid_clicks'),
+            'clicks_this_week'  => (int) DailyEarning::where('user_id', $user->id)->whereBetween('date', [now()->startOfWeek(), now()])->sum('valid_clicks'),
+            'clicks_this_month' => (int) DailyEarning::where('user_id', $user->id)->whereMonth('date', now()->month)->whereYear('date', now()->year)->sum('valid_clicks'),
             'earnings_today'    => $showEarnings ? (float) ($todayEarning?->earnings ?? 0) : null,
             'balance'           => $showEarnings ? (float) ($profile?->balance ?? 0) : null,
             'pending_balance'   => $showEarnings ? (float) ($profile?->pending_balance ?? 0) : null,
